# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.BenefitInfo:
    properties:
      confirmation_subjects:
        items:
          $ref: '#/definitions/dto.ConfirmSubjectResp'
        type: array
      full_score_subjects:
        items:
          type: string
        type: array
      is_bvi:
        type: boolean
      min_class:
        type: integer
      min_diploma_level:
        type: integer
    type: object
  dto.ConfirmSubjectResp:
    properties:
      score:
        type: integer
      subject:
        type: string
    type: object
...
  dto.OlympiadBenefitInfo:
    properties:
      level:
        type: integer
      name:
        type: string
      olympiad_id:
        type: integer
      profile:
        type: string
    type: object
  dto.OlympiadBenefitTree:
    properties:
      benefits:
        items:
          $ref: '#/definitions/dto.BenefitInfo'
        type: array
      olympiad:
        $ref: '#/definitions/dto.OlympiadBenefitInfo'
    type: object
...
  dto.ProgramBenefitInfo:
    properties:
      field:
        type: string
      name:
        type: string
      program_id:
        type: integer
      university:
        type: string
      university_id:
        type: integer
    type: object
  dto.ProgramBenefitTree:
    properties:
      benefits:
        items:
          $ref: '#/definitions/dto.BenefitInfo'
        type: array
      program:
        $ref: '#/definitions/dto.ProgramBenefitInfo'
    type: object
...
  errs.AppError:
    description: Структура ошибки, возвращаемая API в случае неудачного запроса.
    properties:
      code:
        description: HTTP-код ошибки
        type: integer
      details:
        additionalProperties: true
        description: Дополнительные сведения об ошибке (если есть)
        type: object
      message:
        description: Сообщение об ошибке
        type: string
      type:
        description: Тип ошибки
        type: string
    type: object
info:
  contact:
    email: olympguide@mail.ru
    name: Support Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: OlympGuide API
  version: "1.0"
paths:
...
  /olympiads/{id}/benefits:
    get:
      consumes:
      - application/json
      description: |-
        Возвращает список льгот по программам для указанной олимпиады.
        Поддерживаются фильтры по вузу, по направлениям, признаку BVI, минимальному уровню диплома, минимальному классу, а также поиск.
        Льготы сгруппированы по программам, сортировка внутри программы: сначала БВИ, сначала 1 степень.
        Программы сортируются по коду направления подготовки.
      parameters:
      - description: Идентификатор олимпиады
        in: path
        name: id
        required: true
        type: string
      - description: Идентификатор университета (обязательный)
        in: query
        name: university_id
        type: integer
      - collectionFormat: csv
        description: Фильтр по кодам направлений (01.03.04)
        in: query
        items:
          type: string
        name: field
        type: array
      - collectionFormat: csv
        description: Фильтр по BVI
        in: query
        items:
          type: boolean
        name: is_bvi
        type: array
      - collectionFormat: csv
        description: Фильтр по минимальному уровню диплома (1, 2, 3)
        in: query
        items:
          type: integer
        name: min_diploma_level
        type: array
      - collectionFormat: csv
        description: Фильтр по минимальному классу (9, 10, 11)
        in: query
        items:
          type: integer
        name: min_class
        type: array
      - description: Поиск по названию программы и названию университета
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список льгот, сгруппированных по программам
          schema:
            items:
              $ref: '#/definitions/dto.ProgramBenefitTree'
            type: array
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/errs.AppError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: Получить льготы по олимпиаде
      tags:
      - Льготы по олимпиаде
...
  /programs/{id}/benefits:
    get:
      consumes:
      - application/json
      description: |-
        Возвращает список льгот по олимпиадам для указанной образовательной программы.
        Поддерживаются фильтры по уровням олимпиады, профилям олимпиады, признаку BVI, минимальному уровню диплома,
        минимальному классу, а также поиск и сортировка. Льготы сгруппированы по олимпиадам, сортировка внутри программы: сначала БВИ, сначала 1 степень.
      parameters:
      - description: Идентификатор программы
        in: path
        name: id
        required: true
        type: string
      - collectionFormat: csv
        description: Фильтр по уровням олимпиады (1, 2, 3)
        in: query
        items:
          type: string
        name: level
        type: array
      - collectionFormat: csv
        description: Фильтр по профилям (Математика)
        in: query
        items:
          type: string
        name: profile
        type: array
      - collectionFormat: csv
        description: Фильтр по BVI
        in: query
        items:
          type: boolean
        name: is_bvi
        type: array
      - collectionFormat: csv
        description: Фильтр по минимальному уровню диплома (1, 2, 3)
        in: query
        items:
          type: integer
        name: min_diploma_level
        type: array
      - collectionFormat: csv
        description: Фильтр по минимальному классу (9, 10, 11)
        in: query
        items:
          type: integer
        name: min_class
        type: array
      - description: Поиск по названию олимпиады
        in: query
        name: search
        type: string
      - description: Поле сортировки (level - по уровню олимпиады, profile - по профилю
          олимпиады), по умолчанию по убыванию популярности олимпиады.
        in: query
        name: sort
        type: string
      - description: Порядок сортировки (asc или desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список льгот, сгруппированных по олимпиадам
          schema:
            items:
              $ref: '#/definitions/dto.OlympiadBenefitTree'
            type: array
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/errs.AppError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: Получить льготы по программе
      tags:
      - Льготы по программе
...
  /user/diploma/{id}/benefits:
    get:
      consumes:
      - application/json
      description: |-
        Возвращает список льгот по программам для указанного диплома.
        Поддерживаются фильтры по вузу, по направлениям, признаку BVI, минимальному уровню диплома, минимальному классу, а также поиск.
        Льготы сгруппированы по программам, сортировка внутри программы: сначала БВИ, сначала 1 степень.
        Программы сортируются по коду направления подготовки.
      parameters:
      - description: Идентификатор диплома
        in: path
        name: id
        required: true
        type: string
      - description: Идентификатор университета (обязательный)
        in: query
        name: university_id
        type: integer
      - collectionFormat: csv
        description: Фильтр по кодам направлений (01.03.04)
        in: query
        items:
          type: string
        name: field
        type: array
      - collectionFormat: csv
        description: Фильтр по BVI
        in: query
        items:
          type: boolean
        name: is_bvi
        type: array
      - collectionFormat: csv
        description: Фильтр по минимальному уровню диплома (1, 2, 3)
        in: query
        items:
          type: integer
        name: min_diploma_level
        type: array
      - collectionFormat: csv
        description: Фильтр по минимальному классу (9, 10, 11)
        in: query
        items:
          type: integer
        name: min_class
        type: array
      - description: Поиск по названию программы и названию университета
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список льгот, сгруппированных по программам
          schema:
            items:
              $ref: '#/definitions/dto.ProgramBenefitTree'
            type: array
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/errs.AppError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: Получить льготы по диплому
      tags:
      - Льготы по диплому
...
  /user/diplomas/benefits:
    get:
      consumes:
      - application/json
      description: |-
        Возвращает список льгот по программам для пользователя.
        Поддерживаются фильтры по вузу, по направлениям, признаку BVI, минимальному уровню диплома, минимальному классу, а также поиск.
        Льготы сгруппированы по программам, сортировка внутри программы: сначала БВИ, сначала 1 степень.
        Программы сортируются по коду направления подготовки.
      parameters:
      - description: Идентификатор университета (обязательный)
        in: query
        name: university_id
        type: integer
      - collectionFormat: csv
        description: Фильтр по кодам направлений (01.03.04)
        in: query
        items:
          type: string
        name: field
        type: array
      - collectionFormat: csv
        description: Фильтр по BVI
        in: query
        items:
          type: boolean
        name: is_bvi
        type: array
      - collectionFormat: csv
        description: Фильтр по минимальному уровню диплома (1, 2, 3)
        in: query
        items:
          type: integer
        name: min_diploma_level
        type: array
      - collectionFormat: csv
        description: Фильтр по минимальному классу (9, 10, 11)
        in: query
        items:
          type: integer
        name: min_class
        type: array
      - description: Поиск по названию программы и названию университета
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список льгот, сгруппированных по программам
          schema:
            items:
              $ref: '#/definitions/dto.ProgramBenefitTree'
            type: array
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/errs.AppError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: Получить все льготы пользователя
      tags:
      - Льготы пользователя
...
swagger: "2.0"